feat(crowdfundings): Add Membership.needsProlong prop
